requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
discord.py==1.6.0
greenlet==1.0.0
idna==3.1
mariadb==1.0.5
multidict==5.1.0
mysqlclient==2.0.3
SQLAlchemy==1.4.0b3
typing-extensions==3.7.4.3
yarl==1.6.3
